requirements.txt
pytest==2.9.0
Appium-Python-Client
pytest-allure-adaptor